Move ProfilerHook to new contrib.hooks.
Let the new hook bed in in a new place for experimental hooks.
