+++ tensorflow/python/training/basic_session_run_hooks_test.py
@@ -19,7 +19,6 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-import os.path
 import shutil
 import tempfile
 import threading
@@ -39,7 +38,6 @@ from tensorflow.python.ops import state_ops
 from tensorflow.python.ops import variable_scope
 from tensorflow.python.ops import variables as variables_lib
 import tensorflow.python.ops.nn_grad  # pylint: disable=unused-import
-from tensorflow.python.platform import gfile
 from tensorflow.python.platform import test
 from tensorflow.python.platform import tf_logging
 from tensorflow.python.summary import summary as summary_lib
@@ -846,97 +844,6 @@ class FeedFnHookTest(test.TestCase):
       hook.begin()
       mon_sess = monitored_session._HookedSession(sess, [hook])
       self.assertEqual(mon_sess.run(y), 2)
-
-
-class ProfilerHookTest(test.TestCase):
-
-  def setUp(self):
-    super(ProfilerHookTest, self).setUp()
-    self.output_dir = tempfile.mkdtemp()
-    self.graph = ops.Graph()
-    self.filepattern = os.path.join(self.output_dir, "timeline-*.json")
-    with self.graph.as_default():
-      self.global_step = variables.get_or_create_global_step()
-      self.train_op = state_ops.assign_add(self.global_step, 1)
-
-  def tearDown(self):
-    super(ProfilerHookTest, self).tearDown()
-    shutil.rmtree(self.output_dir, ignore_errors=True)
-
-  def countTimelineFiles(self):
-    return len(gfile.Glob(self.filepattern))
-
-  def test_raise_in_both_secs_and_steps(self):
-    with self.assertRaises(ValueError):
-      basic_session_run_hooks.ProfilerHook(save_secs=10, save_steps=20)
-
-  def test_raise_in_none_secs_and_steps(self):
-    with self.assertRaises(ValueError):
-      basic_session_run_hooks.ProfilerHook(save_secs=None, save_steps=None)
-
-  def test_save_secs_saves_in_first_step(self):
-    with self.graph.as_default():
-      hook = basic_session_run_hooks.ProfilerHook(
-          save_secs=2, output_dir=self.output_dir)
-      with monitored_session.SingularMonitoredSession(hooks=[hook]) as sess:
-        sess.run(self.train_op)
-        self.assertEqual(1, self.countTimelineFiles())
-
-  @test.mock.patch('time.time')
-  def test_save_secs_saves_periodically(self, mock_time):
-    # Pick a fixed start time.
-    current_time = 1484863632.320497
-
-    with self.graph.as_default():
-      mock_time.return_value = current_time
-      hook = basic_session_run_hooks.ProfilerHook(
-          save_secs=2, output_dir=self.output_dir)
-      with monitored_session.SingularMonitoredSession(hooks=[hook]) as sess:
-        sess.run(self.train_op)  # Saved.
-        self.assertEqual(1, self.countTimelineFiles())
-        sess.run(self.train_op)  # Not saved.
-        self.assertEqual(1, self.countTimelineFiles())
-        # Simulate 2.5 seconds of sleep.
-        mock_time.return_value = current_time + 2.5
-        sess.run(self.train_op)  # Saved.
-
-        # Pretend some small amount of time has passed.
-        mock_time.return_value = current_time + 0.1
-        sess.run(self.train_op)  # Not saved.
-        # Edge test just before we should save the timeline.
-        mock_time.return_value = current_time + 1.9
-        sess.run(self.train_op)  # Not saved.
-        self.assertEqual(2, self.countTimelineFiles())
-
-        mock_time.return_value = current_time + 4.5
-        sess.run(self.train_op)  # Saved.
-        self.assertEqual(3, self.countTimelineFiles())
-
-  def test_save_steps_saves_in_first_step(self):
-    with self.graph.as_default():
-      hook = basic_session_run_hooks.ProfilerHook(
-          save_secs=2, output_dir=self.output_dir)
-      with monitored_session.SingularMonitoredSession(hooks=[hook]) as sess:
-        sess.run(self.train_op)  # Saved.
-        sess.run(self.train_op)  # Not saved.
-        self.assertEqual(1, self.countTimelineFiles())
-
-  def test_save_steps_saves_periodically(self):
-    with self.graph.as_default():
-      hook = basic_session_run_hooks.ProfilerHook(
-          save_steps=2, output_dir=self.output_dir)
-      with monitored_session.SingularMonitoredSession(hooks=[hook]) as sess:
-        self.assertEqual(0, self.countTimelineFiles())
-        sess.run(self.train_op)  # Saved.
-        self.assertEqual(1, self.countTimelineFiles())
-        sess.run(self.train_op)  # Not saved.
-        self.assertEqual(1, self.countTimelineFiles())
-        sess.run(self.train_op)  # Saved.
-        self.assertEqual(2, self.countTimelineFiles())
-        sess.run(self.train_op)  # Not saved.
-        self.assertEqual(2, self.countTimelineFiles())
-        sess.run(self.train_op)  # Saved.
-        self.assertEqual(3, self.countTimelineFiles())
 
 
 if __name__ == '__main__':
