+++ tensorflow/python/training/training.py
@@ -141,7 +141,6 @@ Hooks are tools that run in the process of training/evaluation of the model.
 @@GlobalStepWaiterHook
 @@FinalOpsHook
 @@FeedFnHook
-@@ProfilerHook
 
 ## Training Utilities
 
@@ -202,7 +201,7 @@ from tensorflow.python.training.basic_session_run_hooks import SummarySaverHook
 from tensorflow.python.training.basic_session_run_hooks import GlobalStepWaiterHook
 from tensorflow.python.training.basic_session_run_hooks import FinalOpsHook
 from tensorflow.python.training.basic_session_run_hooks import FeedFnHook
-from tensorflow.python.training.basic_session_run_hooks import ProfilerHook
+
 from tensorflow.python.training.basic_loops import basic_train_loop
 from tensorflow.python.training.device_setter import replica_device_setter
 from tensorflow.python.training.monitored_session import Scaffold
